Added supp.evidence to Obj->write_Gene and to update.pl Added also -nowrite -usefile and timdb to update.pl

+++ scripts/update.pl
@@ -30,9 +30,12 @@ This script updates a recipient database by checking its donor database
     -arcpass   password for the archive database
 
     -help      Displays script documentation with PERLDOC
+    
+    -nowrite   Runs entire script without writing in recipient
 
     -verbose   Gets all the print STDERR for testing purposes
 
+    -usefile   read in on stdin a list of clones, one clone per line
 =cut
 
 use strict;
@@ -52,25 +55,41 @@ my $dbuser = 'root';
 my $dbpass = undef;
 my $module = 'Bio::EnsEMBL::DBSQL::Obj';
 my $help;
+my $nowrite;
 my $arcpass = undef;
 my $verbose;
+my $usefile = 0;
+
 
 &GetOptions( 
-	     'dbtype:s'   => \$dbtype,
-	     'host:s'     => \$host,
-	     'port:n'     => \$port,
-	     'dbname:s'   => \$dbname,
-	     'dbuser:s'   => \$dbuser,
-	     'dbpass:s'   => \$dbpass,
-	     'module:s'   => \$module,
-	     'arcpass:s'  => \$arcpass,
-	     'h|help'     => \$help,
-	     'v|verbose'  => \$verbose
+	     'dbtype:s'  => \$dbtype,
+	     'host:s'    => \$host,
+	     'port:n'    => \$port,
+	     'dbname:s'  => \$dbname,
+	     'dbuser:s'  => \$dbuser,
+	     'dbpass:s'  => \$dbpass,
+	     'module:s'  => \$module,
+	     'arcpass:s' => \$arcpass,
+	     'h|help'    => \$help,
+	     'nowrite'   => \$nowrite,
+	     'usefile'   => \$usefile,
+	     'v|verbose' => \$verbose
 	     );
 
 
 if ($help) {
     exec('perldoc', $0);
+}
+
+my $don_db;
+my @clones; 
+my @object_array;
+
+if( $usefile == 1 ) {
+    while( <> ) {
+	my ($en) = split;
+	push(@clones,$en);
+    }
 }
 
 my $arcname = 'archive';
@@ -92,8 +111,13 @@ my $rec_db =  Bio::EnsEMBL::DBLoader->new($recipient_locator);
 
 $verbose && print "\nConnecting to donor database...\n";
 my $donor_locator = $rec_db->get_donor_locator;
-my $don_db =  Bio::EnsEMBL::DBLoader->new($donor_locator);
 
+if ($donor_locator eq 'timdb') {
+    $don_db = Bio::EnsEMBL::TimDB::Obj->new(\@clones); 
+}
+else {
+    $don_db =  Bio::EnsEMBL::DBLoader->new($donor_locator);
+}
 my $last_offset = $rec_db->get_last_update_offset;
 $verbose && print "\nLast update-offset: $last_offset\n";
 
@@ -105,47 +129,49 @@ if ($last_offset > $now_offset) {
     exit;
 }
 
-$verbose && print "\nTransferring new ghosts from donor to recipient...\n";
+if ($donor_locator ne 'timdb') {
+    $verbose && print "\nTransferring new ghosts from donor to recipient...\n";
 
-my @object_array = $don_db->get_updated_Ghosts($last_offset, $now_offset);
-
-#Sort with genes before exons (hopefully!)
-@object_array = sort { if ( $a->obj_type eq 'gene' ) { return 1; } else { return $a->obj_type cmp $b->obj_type } } @object_array;
-
-foreach my $ghost (@object_array) {
-    $verbose && print "Got ghost with id ".$ghost->id."\n";
+    @object_array = $don_db->get_updated_Ghosts($last_offset, $now_offset);
     
-    #Check if the ghost is already present in the recipient database
-    eval {
-	$rec_db->get_Ghost($ghost->id,$ghost->obj_type);
-    };
+    #Sort with genes before exons (hopefully!)
+    @object_array = sort { if ( $a->obj_type eq 'gene' ) { return 1; } else { return $a->obj_type cmp $b->obj_type } } @object_array;
     
-    #If not present in recipient, archive objects, and write the ghost
-    if ( $@ ) {
-	#If ghost of clone, delete clone
-	if ($ghost->obj_type eq 'clone') {
-	    $verbose && print "New ghost for deleted clone ".$ghost->id.", deleting clone from database\n";
-	    $rec_db->delete_Clone($ghost->id);
+    foreach my $ghost (@object_array) {
+	$verbose && print "Got ghost with id ".$ghost->id."\n";
+	
+	#Check if the ghost is already present in the recipient database
+	eval {
+	    $rec_db->get_Ghost($ghost->id,$ghost->obj_type);
+	};
+	
+	#If not present in recipient, archive objects, and write the ghost
+	if ( $@ ) {
+	    #If ghost of clone, delete clone
+	    if ($ghost->obj_type eq 'clone') {
+		$verbose && print "New ghost for deleted clone ".$ghost->id.", deleting clone from database\n";
+		$nowrite || $rec_db->delete_Clone($ghost->id);
+	    }
+	    #If ghost of gene, archive gene
+	    elsif ($ghost->obj_type eq 'gene') {
+		$verbose && print "New ghost for deleted gene ".$ghost->id.", archiving gene from recipient to archive\n";
+		my $gene = $rec_db->get_Gene($ghost->id);
+		$nowrite || $rec_db->archive_Gene($gene,$arc_db);
+	    }
+	    #If ghost of exon, delete exon
+	    elsif ($ghost->obj_type eq 'exon') {
+		$verbose && print "New ghost for deleted exon ".$ghost->id.", deleting exon from database\n";
+		$nowrite || $rec_db->delete_Exon($ghost->id);
+	    }
+	    #Finally write ghost in recipient
+	    $verbose && print "Writing ghost in recipient database\n";
+	    $nowrite || $rec_db->write_Ghost($ghost);
 	}
-	#If ghost of gene, archive gene
-	elsif ($ghost->obj_type eq 'gene') {
-	    $verbose && print "New ghost for deleted gene ".$ghost->id.", archiving gene from recipient to archive\n";
-	    my $gene = $rec_db->get_Gene($ghost->id);
-	    $rec_db->archive_Gene($gene,$arc_db);
-	}
-	#If ghost of exon, delete exon
-	elsif ($ghost->obj_type eq 'exon') {
-            $verbose && print "New ghost for deleted exon ".$ghost->id.", deleting exon from database\n";
-	    $rec_db->delete_Exon($ghost->id);
-	}
-	#Finally write ghost in recipient
-	$verbose && print "Writing ghost in recipient database\n";
-	$rec_db->write_Ghost($ghost);
+	
+	#Nothing needs to be done if the ghost is already present
     }
-
-    #Nothing needs to be done if the ghost is already present
 }
-
+    
 $verbose && print "\nTransferring updated and new objects from donor to recipient...\n";
 #Get updated and new objects (clones and genes)
 @object_array = [];
@@ -168,10 +194,10 @@ foreach my $object (@object_array) {
         #If not present in recipient, write the clone in, and get all its genes out
 	if ( $@ ) { 
 	    $verbose &&  print "New Clone, writing it in the database\n";
-	    $rec_db->write_Clone($object);
+	    $nowrite || $rec_db->write_Clone($object);
 	    
 	    #Get all the genes from this clone and check them, write them, archive them accordingly
-	    foreach my $gene ($object->get_all_Genes()) {
+	    foreach my $gene ($object->get_all_Genes('evidence')) {
 		$verbose &&  print "Getting all genes via clone get_all_Genes method\n";
 		&_place_gene($gene,'1');
 	    }
@@ -183,10 +209,10 @@ foreach my $object (@object_array) {
 	    #NOTE: We permanently delete clones, without archiving them!
 	    if ($object->version > $rec_clone->version) {
 		$verbose && print "Clone with new version, updating the database, and deleting the old version\n";
-		$rec_db->delete_Clone($rec_clone->id);
-		$rec_db->write_Clone($object);
+		$nowrite || $rec_db->delete_Clone($rec_clone->id);
+		$nowrite || $rec_db->write_Clone($object);
 		#Get all the genes from this clone
-		foreach my $gene ($object->get_all_Genes()) {
+		foreach my $gene ($object->get_all_Genes('evidence')) {
 		    $verbose &&  print "Getting all genes via clone get_all_Genes method\n";
 		    $verbose &&  print "CLONE LEVEL: Got gene ".$gene->id."\n";
 		    &_place_gene($gene,'1');
@@ -226,15 +252,15 @@ foreach my $object (@object_array) {
 	#If not present in recipient, write it in
 	if ( $@ ) {
 	    $verbose && print "New Exon, writing it in the database\n";
-	    $rec_db->write_Exon($object);
+	    $nowrite || $rec_db->write_Exon($object);
 	}
 	#If exon present in recipient, check donor and recipient version
 	else {
 	#If donor exon version greater than recipient exon version, update
 	    if ($object->version > $rec_exon->version) {
 		$verbose && print "Exon with new version, updating the database\n";
-		$rec_db->delete_Exon($object->id);
-		$rec_db->write_Exon($object);
+		$nowrite || $rec_db->delete_Exon($object->id);
+		$nowrite || $rec_db->write_Exon($object);
 	    }
 	
 	    #If donor gene version is less than the recipient gene version, error 
@@ -255,7 +281,7 @@ foreach my $object (@object_array) {
 }
 
 
-$rec_db->replace_last_update($now_offset);
+$nowrite || $rec_db->replace_last_update($now_offset);
 
 sub _place_gene {
     my ($don_gene,$clone_level) = @_;
@@ -263,22 +289,22 @@ sub _place_gene {
 
     #Check if the gene is present in the recipient
     eval {
-	$rec_gene = $rec_db->get_Gene($don_gene->id);
+	$rec_gene = $rec_db->get_Gene($don_gene->id,'evidence');
     };
     
     #If gene not present in recipient, write it in
     if ( $@ ) {
 	$verbose && print "New Gene, writing it in the database\n";
 	$verbose &&  print "In _place_gene: Gene ".$don_gene->id." has version ".$don_gene->version."\n";
-	$rec_db->write_Gene($don_gene);
+	$nowrite || $rec_db->write_Gene($don_gene,'evidence');
     }
     #If gene present in recipient, check donor and recipient version
     else {
 	#If donor gene version greater than recipient gene version, update
 	if ($don_gene->version > $rec_gene->version) {
 	    $verbose && print "Gene with new version, updating the database, and archiving old version\n";
-	    $rec_db->archive_Gene($rec_gene,$arc_db);
-	    $rec_db->write_Gene($don_gene);
+	    $nowrite || $rec_db->archive_Gene($rec_gene,$arc_db);
+	    $nowrite || $rec_db->write_Gene($don_gene,'evidence');
 	}
 	
 	#If donor gene version is less than the recipient gene version, error 
@@ -290,8 +316,8 @@ sub _place_gene {
 	else {
 	    if ($clone_level) {
 		$verbose && print "Genes with the same version, deleting recipient gene and writing one from donor without archiving\n";  
-		$rec_db->delete_Gene($rec_gene->id);
-		$rec_db->write_Gene($don_gene);
+		$nowrite || $rec_db->delete_Gene($rec_gene->id);
+		$nowrite || $rec_db->write_Gene($don_gene,'evidence');
 	    }
 	    else {
 		$verbose && print "Genes with the same version, nothing needs to be done\n";
